automatically backed up by learn

+++ app/controllers/application_controller.rb
@@ -66,8 +66,13 @@ class ApplicationController < Sinatra::Base
 
   get '/tweets/:id/edit' do
     tweet = Tweet.find_by(id: params[:id])
-    tweet.destroy
-    "Tweet was deleted"
+    user = User.find_by(id: session[:user_id])
+    if user.id == tweet.user.id
+      tweet.destroy
+      "Tweet was deleted"
+    else
+      "Can't delete someone else's tweet"
+    end
   end
 
   post '/signup' do
